All Script
This script is designed to run in Google Colaboratory. If you wish to apply it to other Jupyter Notebooks or platforms, you may need to adjust parameters and syntax according to the specific requirements of the environment. However, the script itself should work seamlessly in Google Colab with minimal modifications.